Add a secondary context path
What if we wanted to host the web site at another context path instead of
the root location? We can change that through the getServletMappings
method. You can have multiple servlet mappings (such as in this case) or
remove the root mapping entirely and just have a single one.

Yes, the asterisk is important. It tells Spring to use everything after
/example/ for routing. That way you do not have to start all of your URLs
off with '/example/' (i.e., '/example/index').